update webpack.common.js
Added resolve to allow import of new js/jsx components. Also i had tried to figure out how to avoid using of the constructor to declare a state and just declare the state as is, but still getting some errors. I finally use the constructors and could built the component, hence the project built just fine.